+++ src/main/java/uk/q3c/krail/core/data/AnnotationOptionListConverter.java
@@ -16,14 +16,14 @@ package uk.q3c.krail.core.data;
 import com.google.common.base.Splitter;
 import com.google.common.collect.ImmutableList;
 import org.apache.commons.lang3.StringEscapeUtils;
-import uk.q3c.krail.core.option.AnnotationOptionList;
+import uk.q3c.util.collection.AnnotationList;
 
 import javax.annotation.Nonnull;
 import java.lang.annotation.Annotation;
 import java.util.ArrayList;
 import java.util.List;
 
-import static com.google.common.base.Preconditions.checkNotNull;
+import static com.google.common.base.Preconditions.*;
 
 
 /**
@@ -46,7 +46,7 @@ public class AnnotationOptionListConverter {
      * @param model the model to convert
      * @return Comma separated String representation of the list contents
      */
-    public String convertToString(@Nonnull AnnotationOptionList model) {
+    public String convertToString(@Nonnull AnnotationList model) {
         checkNotNull(model);
         if (model.isEmpty()) {
             return "";
@@ -73,10 +73,10 @@ public class AnnotationOptionListConverter {
      * @return Comma separated String representation of the list contents
      * @throws ConversionException if conversion fails
      */
-    public AnnotationOptionList convertToModel(@Nonnull String value) {
+    public AnnotationList convertToModel(@Nonnull String value) {
         checkNotNull(value);
         if (value.isEmpty()) {
-            return new AnnotationOptionList();
+            return new AnnotationList();
         }
         final List<String> strings = Splitter.on(separator)
                                              .splitToList(value);
@@ -91,6 +91,6 @@ public class AnnotationOptionListConverter {
                 throw new ConversionException("Class " + element + " is not an Annotation");
             }
         });
-        return new AnnotationOptionList(elementList);
+        return new AnnotationList(elementList);
     }
 }