+++ src/main/java/uk/q3c/krail/core/data/DefaultOptionElementConverter.java
@@ -17,10 +17,10 @@ import com.google.common.primitives.Ints;
 import com.google.common.primitives.Longs;
 import com.google.inject.Inject;
 import com.vaadin.data.util.converter.ConverterFactory;
-import uk.q3c.krail.core.option.AnnotationOptionList;
 import uk.q3c.krail.core.option.OptionList;
 import uk.q3c.krail.i18n.api.I18NKey;
 import uk.q3c.krail.i18n.api.MessageFormat2;
+import uk.q3c.util.collection.AnnotationList;
 
 import javax.annotation.Nonnull;
 import java.math.BigDecimal;
@@ -74,8 +74,8 @@ public class DefaultOptionElementConverter implements OptionElementConverter {
             return value.toString();
         } else if (OptionList.class.isAssignableFrom(modelType)) {
             return new OptionListConverter(this).convertToString((OptionList) value);
-        } else if (AnnotationOptionList.class.isAssignableFrom(modelType)) {
-            return new AnnotationOptionListConverter().convertToString((AnnotationOptionList) value);
+        } else if (AnnotationList.class.isAssignableFrom(modelType)) {
+            return new AnnotationOptionListConverter().convertToString((AnnotationList) value);
         }
         String msg = messageFormat.format("Data type of {0} is not supported in Option", value.getClass());
         throw new ConverterException(msg);
@@ -106,7 +106,7 @@ public class DefaultOptionElementConverter implements OptionElementConverter {
             return (V) new EnumConverter().convertToModel(valueString);
         } else if (elementClass == BigDecimal.class) {
             return (V) new BigDecimal(valueString);
-        } else if (elementClass == AnnotationOptionList.class) {
+        } else if (elementClass == AnnotationList.class) {
             return (V) new AnnotationOptionListConverter().convertToModel(valueString);
         }
         String msg = messageFormat.format("Data type of {0} is not supported in Option", elementClass);