+++ src/main/java/uk/q3c/krail/i18n/PatternSourceProviderConfigByOption.kt
@@ -3,30 +3,30 @@ package uk.q3c.krail.i18n
 import com.google.inject.Inject
 import uk.q3c.krail.core.i18n.DescriptionKey
 import uk.q3c.krail.core.i18n.LabelKey
-import uk.q3c.krail.core.option.AnnotationOptionList
 import uk.q3c.krail.core.option.Option
 import uk.q3c.krail.core.option.OptionContext
 import uk.q3c.krail.core.option.OptionKey
 import uk.q3c.krail.i18n.api.PatternSourceProvider
 import uk.q3c.krail.i18n.api.PatternSourceProviderConfig
+import uk.q3c.util.collection.AnnotationList
 
 /**
  * Created by David Sowerby on 04 Aug 2017
  */
 class PatternSourceProviderConfigByOption @Inject constructor(val option: Option) : PatternSourceProviderConfig, OptionContext<Any> {
 
-    val optionKeySourceOrder: OptionKey<AnnotationOptionList> =
-            OptionKey(AnnotationOptionList(), PatternSourceProviderConfigByOption::class.java, LabelKey.Source_Order, DescriptionKey.Source_Order)
+    val optionKeySourceOrder: OptionKey<AnnotationList> =
+            OptionKey(AnnotationList(), PatternSourceProviderConfigByOption::class.java, LabelKey.Source_Order, DescriptionKey.Source_Order)
 
-    val optionKeySourceOrderDefault: OptionKey<AnnotationOptionList> = OptionKey(AnnotationOptionList(), PatternSourceProviderConfigByOption::class.java,
+    val optionKeySourceOrderDefault: OptionKey<AnnotationList> = OptionKey(AnnotationList(), PatternSourceProviderConfigByOption::class.java,
             LabelKey.Source_Order_Default, DescriptionKey.Source_Order_Default)
 
-    val optionKeySelectedTargets: OptionKey<AnnotationOptionList> = OptionKey(AnnotationOptionList(), PatternSourceProviderConfigByOption::class.java,
+    val optionKeySelectedTargets: OptionKey<AnnotationList> = OptionKey(AnnotationList(), PatternSourceProviderConfigByOption::class.java,
             LabelKey.Selected_Pattern_Targets, DescriptionKey.Selected_Pattern_Targets)
 
     override lateinit var patternSourceProvider: PatternSourceProvider
 
-    override var sourceOrder: AnnotationOptionList = AnnotationOptionList()
+    override var sourceOrder: AnnotationList = AnnotationList()
         get() {
             return option[optionKeySourceOrder]
         }
@@ -36,7 +36,7 @@ class PatternSourceProviderConfigByOption @Inject constructor(val option: Option
         }
 
 
-    override var sourceOrderDefault: AnnotationOptionList = AnnotationOptionList()
+    override var sourceOrderDefault: AnnotationList = AnnotationList()
         get() {
             return option[optionKeySourceOrderDefault]
         }
@@ -45,7 +45,7 @@ class PatternSourceProviderConfigByOption @Inject constructor(val option: Option
             option.set(optionKeySourceOrderDefault, value)
         }
 
-    override var selectedTargets: AnnotationOptionList = AnnotationOptionList()
+    override var selectedTargets: AnnotationList = AnnotationList()
         get() {
             return option[optionKeySelectedTargets]
         }